Restructure Shared Code
