+++ mock/local-journal.json
@@ -9,7 +9,7 @@
       "slotDetail": {
         "duration": 57,
         "slotId": 66558326,
-        "start": "2019-02-21T15:54:00+00:00"
+        "start": "2019-02-26T15:54:00+00:00"
       },
       "testCentre": {
         "centreId": 1878,
@@ -81,7 +81,7 @@
       "slotDetail": {
         "duration": 57,
         "slotId": 66558319,
-        "start": "2019-02-21T08:40:00+00:00"
+        "start": "2019-02-26T08:40:00+00:00"
       },
       "testCentre": {
         "centreId": 1878,
@@ -152,7 +152,7 @@
       "slotDetail": {
         "duration": 57,
         "slotId": 66558320,
-        "start": "2019-02-21T09:37:00+00:00"
+        "start": "2019-02-26T09:37:00+00:00"
       },
       "testCentre": {
         "centreId": 1878,
@@ -223,7 +223,7 @@
       "slotDetail": {
         "duration": 57,
         "slotId": 66558321,
-        "start": "2019-02-21T10:44:00+00:00"
+        "start": "2019-02-26T10:44:00+00:00"
       },
       "testCentre": {
         "centreId": 1878,
@@ -294,7 +294,7 @@
       "slotDetail": {
         "duration": 57,
         "slotId": 66558322,
-        "start": "2019-02-21T11:41:00+00:00"
+        "start": "2019-02-26T11:41:00+00:00"
       },
       "testCentre": {
         "centreId": 1878,
@@ -365,7 +365,7 @@
       "slotDetail": {
         "duration": 57,
         "slotId": 66558323,
-        "start": "2019-02-21T13:03:00+00:00"
+        "start": "2019-02-26T13:03:00+00:00"
       },
       "testCentre": {
         "centreId": 1878,
@@ -436,7 +436,7 @@
       "slotDetail": {
         "duration": 57,
         "slotId": 66558324,
-        "start": "2019-02-21T14:00:00+00:00"
+        "start": "2019-02-26T14:00:00+00:00"
       },
       "testCentre": {
         "centreId": 1878,
@@ -507,7 +507,7 @@
       "slotDetail": {
         "duration": 57,
         "slotId": 66558325,
-        "start": "2019-02-21T14:57:00+00:00"
+        "start": "2019-02-26T14:57:00+00:00"
       },
       "testCentre": {
         "centreId": 1878,
@@ -578,7 +578,7 @@
       "slotDetail": {
         "duration": 57,
         "slotId": 66558354,
-        "start": "2019-02-22T08:40:00+00:00"
+        "start": "2019-02-27T08:40:00+00:00"
       },
       "testCentre": {
         "centreId": 1878,
@@ -649,7 +649,7 @@
       "slotDetail": {
         "duration": 57,
         "slotId": 66558355,
-        "start": "2019-02-22T09:37:00+00:00"
+        "start": "2019-02-27T09:37:00+00:00"
       },
       "testCentre": {
         "centreId": 1878,
@@ -720,7 +720,7 @@
       "slotDetail": {
         "duration": 57,
         "slotId": 66558356,
-        "start": "2019-02-22T10:44:00+00:00"
+        "start": "2019-02-27T10:44:00+00:00"
       },
       "testCentre": {
         "centreId": 1878,
@@ -791,7 +791,7 @@
       "slotDetail": {
         "duration": 57,
         "slotId": 66558357,
-        "start": "2019-02-22T11:41:00+00:00"
+        "start": "2019-02-27T11:41:00+00:00"
       },
       "testCentre": {
         "centreId": 1878,
@@ -862,7 +862,7 @@
       "slotDetail": {
         "duration": 57,
         "slotId": 66558358,
-        "start": "2019-02-22T13:03:00+00:00"
+        "start": "2019-02-27T13:03:00+00:00"
       },
       "testCentre": {
         "centreId": 1878,
@@ -933,7 +933,7 @@
       "slotDetail": {
         "duration": 57,
         "slotId": 66558359,
-        "start": "2019-02-22T14:00:00+00:00"
+        "start": "2019-02-27T14:00:00+00:00"
       },
       "testCentre": {
         "centreId": 1878,
@@ -1004,7 +1004,7 @@
       "slotDetail": {
         "duration": 57,
         "slotId": 66558360,
-        "start": "2019-02-22T14:57:00+00:00"
+        "start": "2019-02-27T14:57:00+00:00"
       },
       "testCentre": {
         "centreId": 1878,
