+++ src/app/app.component.ts
@@ -3,7 +3,7 @@ import { Platform } from 'ionic-angular';
 import { StatusBar } from '@ionic-native/status-bar';
 import { Store } from '@ngrx/store';
 
-import { StoreModel } from '../common/store.model';
+import { StoreModel } from '../shared/models/store.model';
 import { LoadAppInfo } from '../modules/app-info/app-info.actions';
 
 declare let window: any;
