+++ src/pages/candidate-details/candidate-details.ts
@@ -5,10 +5,10 @@ import { Observable } from 'rxjs/Observable';
 import { Subscription } from 'rxjs/Subscription';
 import { map } from 'rxjs/operators';
 import { merge } from 'rxjs/observable/merge';
-import { BasePageComponent } from '../../classes/base-page';
+import { BasePageComponent } from '../../shared/classes/base-page';
 import { AuthenticationProvider } from '../../providers/authentication/authentication';
-import { StoreModel } from '../../common/store.model';
-import { TestCategory, testCategoryIcons } from '../../common/test-category';
+import { StoreModel } from '../../shared/models/store.model';
+import { TestCategory, testCategoryIcons } from '../../shared/models/test-category';
 import { getJournalState } from '../journal/journal.reducer';
 import { Details } from './candidate-details.model';
 import { ClearChangedSlot } from '../journal/journal.actions';
