+++ src/pages/journal/components/journal-navigation/__mocks__/journal-slots-data.mock.ts
@@ -1,5 +1,5 @@
 import { SlotItem } from '../../../../../providers/slot-selector/slot-item';
-import { DateTime, Duration } from '../../../../../common/date-time';
+import { DateTime, Duration } from '../../../../../shared/helpers/date-time';
 
 // In order for the navigation to work we just need 3 days in slots
 // The days can be empty because we don't care about the actual slots
