+++ src/pages/journal/journal.model.ts
@@ -1,4 +1,4 @@
-import { MesError } from '../../common/mes-error.model';
+import { MesError } from '../../shared/models/mes-error.model';
 import { SlotItem } from '../../providers/slot-selector/slot-item';
 
 export type Slot = {
