+++ src/shared/models/DJournal.ts
