+++ src/shared/models/store.model.ts
@@ -1,0 +1,7 @@
+import { JournalModel } from '../../pages/journal/journal.model';
+import { AppInfoModel } from '../../modules/app-info/app-info.model';
+
+export type StoreModel = {
+  journal: JournalModel,
+  appInfo: AppInfoModel,
+};
